backend/src/main/python/requirements.txt
flask
torch
torchvision
pillow
ftfy
regex
tqdm
